♻️ Refactor name of SubmitButton
