+++ src/app/components/Buttons/SubmitButton/SubmitButton.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import AddIcon from '../../assets/AddIcon';
 import SearchIcon from '../../assets/SearchIcon';
-import InputButton from './InputButton';
+import InputButton from './SubmitButton';
 
 export default {
   title: 'Component/Button',
